+++ sdk/keyvault/azure-keyvault-certificates/tests/test_merge_certificate_async.py
@@ -10,13 +10,14 @@ from azure.keyvault.certificates.aio import CertificateClient
 from devtools_testutils import ResourceGroupPreparer, KeyVaultPreparer
 from OpenSSL import crypto
 
+from _shared.json_attribute_matcher import json_attribute_matcher
 from _shared.preparer_async import KeyVaultClientPreparer
 from _shared.test_case_async import KeyVaultTestCase
 
 
 class MergeCertificateTest(KeyVaultTestCase):
     def __init__(self, *args, **kwargs):
-        super().__init__(*args, match_body=False, **kwargs)
+        super().__init__(*args, match_body=False, custom_request_matchers=[json_attribute_matcher], **kwargs)
 
     @ResourceGroupPreparer(random_name_enabled=True)
     @KeyVaultPreparer()