+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/_models.py
@@ -391,10 +391,10 @@ class AnalyzeSentimentResult(DictMixin):
         field will contain information about the document payload.
     :type statistics:
         ~azure.ai.textanalytics.TextDocumentStatistics
-    :param sentiment_scores: Document level sentiment confidence
-        scores between 0 and 1 for each sentiment class.
-    :type sentiment_scores:
-        ~azure.ai.textanalytics.SentimentScorePerLabel
+    :param confidence_scores: Document level sentiment confidence
+        scores between 0 and 1 for each sentiment label.
+    :type confidence_scores:
+        ~azure.ai.textanalytics.SentimentConfidenceScorePerLabel
     :param sentences: Sentence level sentiment analysis.
     :type sentences:
         list[~azure.ai.textanalytics.SentenceSentiment]
@@ -406,13 +406,13 @@ class AnalyzeSentimentResult(DictMixin):
         self.id = kwargs.get("id", None)
         self.sentiment = kwargs.get("sentiment", None)
         self.statistics = kwargs.get("statistics", None)
-        self.sentiment_scores = kwargs.get("sentiment_scores", None)
+        self.confidence_scores = kwargs.get("confidence_scores", None)
         self.sentences = kwargs.get("sentences", None)
         self.is_error = False
 
     def __repr__(self):
-        return "AnalyzeSentimentResult(id={}, sentiment={}, statistics={}, sentiment_scores={}, sentences={}, " \
-               "is_error={})".format(self.id, self.sentiment, repr(self.statistics), repr(self.sentiment_scores),
+        return "AnalyzeSentimentResult(id={}, sentiment={}, statistics={}, confidence_scores={}, sentences={}, " \
+               "is_error={})".format(self.id, self.sentiment, repr(self.statistics), repr(self.confidence_scores),
                                      repr(self.sentences), self.is_error)[:1024]
 
 
@@ -532,9 +532,9 @@ class LinkedEntity(DictMixin):
         list[~azure.ai.textanalytics.LinkedEntityMatch]
     :param language: Language used in the data source.
     :type language: str
-    :param id: Unique identifier of the recognized entity from the data
+    :param data_source_entity_id: Unique identifier of the recognized entity from the data
         source.
-    :type id: str
+    :type data_source_entity_id: str
     :param url: URL to the entity's page from the data source.
     :type url: str
     :param data_source: Data source used to extract entity linking,
@@ -546,7 +546,7 @@ class LinkedEntity(DictMixin):
         self.name = kwargs.get("name", None)
         self.matches = kwargs.get("matches", None)
         self.language = kwargs.get("language", None)
-        self.id = kwargs.get("id", None)
+        self.data_source_entity_id = kwargs.get("data_source_entity_id", None)
         self.url = kwargs.get("url", None)
         self.data_source = kwargs.get("data_source", None)
 
@@ -556,15 +556,15 @@ class LinkedEntity(DictMixin):
             name=entity.name,
             matches=[LinkedEntityMatch._from_generated(e) for e in entity.matches],  # pylint: disable=protected-access
             language=entity.language,
-            id=entity.id,
+            data_source_entity_id=entity.id,
             url=entity.url,
             data_source=entity.data_source,
         )
 
     def __repr__(self):
-        return "LinkedEntity(name={}, matches={}, language={}, id={}, url={}, " \
-               "data_source={})".format(self.name, repr(self.matches), self.language, self.id, self.url,
-                                        self.data_source)[:1024]
+        return "LinkedEntity(name={}, matches={}, language={}, data_source_entity_id={}, url={}, " \
+               "data_source={})".format(self.name, repr(self.matches), self.language, self.data_source_entity_id,
+                                        self.url, self.data_source)[:1024]
 
 
 class LinkedEntityMatch(DictMixin):
@@ -674,10 +674,10 @@ class SentenceSentiment(DictMixin):
     :param sentiment: The predicted Sentiment for the sentence.
         Possible values include: 'positive', 'neutral', 'negative'
     :type sentiment: str
-    :param sentiment_scores: The sentiment confidence score between 0
+    :param confidence_scores: The sentiment confidence score between 0
         and 1 for the sentence for all labels.
-    :type sentiment_scores:
-        ~azure.ai.textanalytics.SentimentScorePerLabel
+    :type confidence_scores:
+        ~azure.ai.textanalytics.SentimentConfidenceScorePerLabel
     :param offset: The sentence offset from the start of the
         document.
     :type offset: int
@@ -689,7 +689,7 @@ class SentenceSentiment(DictMixin):
 
     def __init__(self, **kwargs):
         self.sentiment = kwargs.get("sentiment", None)
-        self.sentiment_scores = kwargs.get("sentiment_scores", None)
+        self.confidence_scores = kwargs.get("confidence_scores", None)
         self.offset = kwargs.get("offset", None)
         self.length = kwargs.get("length", None)
         self.warnings = kwargs.get("warnings", None)
@@ -698,20 +698,20 @@ class SentenceSentiment(DictMixin):
     def _from_generated(cls, sentence):
         return cls(
             sentiment=sentence.sentiment.value,
-            sentiment_scores=SentimentScorePerLabel._from_generated(sentence.sentence_scores),  # pylint: disable=protected-access
+            confidence_scores=SentimentConfidenceScorePerLabel._from_generated(sentence.sentence_scores),  # pylint: disable=protected-access
             offset=sentence.offset,
             length=sentence.length,
             warnings=sentence.warnings,
         )
 
     def __repr__(self):
-        return "SentenceSentiment(sentiment={}, sentiment_scores={}, offset={}, length={}, warnings={})" \
-            .format(self.sentiment, repr(self.sentiment_scores), self.offset, self.length, self.warnings)[:1024]
+        return "SentenceSentiment(sentiment={}, confidence_scores={}, offset={}, length={}, warnings={})" \
+            .format(self.sentiment, repr(self.confidence_scores), self.offset, self.length, self.warnings)[:1024]
 
 
-class SentimentScorePerLabel(DictMixin):
-    """Represents the confidence scores between 0 and 1 across all sentiment
-    labels: positive, neutral, negative.
+class SentimentConfidenceScorePerLabel(DictMixin):
+    """The confidence scores (Softmax scores) between 0 and 1 across all sentiment
+    labels: positive, neutral, negative. Higher values indicate higher confidence.
 
     :param positive: Positive score.
     :type positive: float
@@ -735,5 +735,5 @@ class SentimentScorePerLabel(DictMixin):
         )
 
     def __repr__(self):
-        return "SentimentScorePerLabel(positive={}, neutral={}, negative={})" \
+        return "SentimentConfidenceScorePerLabel(positive={}, neutral={}, negative={})" \
             .format(self.positive, self.neutral, self.negative)[:1024]