+++ sdk/textanalytics/azure-ai-textanalytics/tests/test_text_analytics.py
@@ -74,7 +74,7 @@ class TextAnalyticsTest(TestAnalyticsTestCase):
             name="Bill Gates",
             matches=[linked_entity_match, linked_entity_match],
             language="English",
-            id="Bill Gates",
+            data_source_entity_id="Bill Gates",
             url="https://en.wikipedia.org/wiki/Bill_Gates",
             data_source="wikipedia"
         )
@@ -84,11 +84,11 @@ class TextAnalyticsTest(TestAnalyticsTestCase):
             )
 
         sentiment_confidence_score_per_label = \
-            _models.SentimentScorePerLabel(positive=0.99, neutral=0.05, negative=0.02)
+            _models.SentimentConfidenceScorePerLabel(positive=0.99, neutral=0.05, negative=0.02)
 
         sentence_sentiment = _models.SentenceSentiment(
             sentiment="neutral",
-            sentiment_scores=sentiment_confidence_score_per_label,
+            confidence_scores=sentiment_confidence_score_per_label,
             offset=0,
             length=10,
             warnings=["sentence was too short to find sentiment"]
@@ -98,7 +98,7 @@ class TextAnalyticsTest(TestAnalyticsTestCase):
             id="1",
             sentiment="positive",
             statistics=text_document_statistics,
-            sentiment_scores=sentiment_confidence_score_per_label,
+            confidence_scores=sentiment_confidence_score_per_label,
             sentences=[sentence_sentiment],
             is_error=False
         )
@@ -142,23 +142,23 @@ class TextAnalyticsTest(TestAnalyticsTestCase):
         self.assertEqual("LinkedEntityMatch(score=0.999, text=Bill Gates, offset=0, length=8)", repr(linked_entity_match))
         self.assertEqual("LinkedEntity(name=Bill Gates, matches=[LinkedEntityMatch(score=0.999, text=Bill Gates, "
                          "offset=0, length=8), LinkedEntityMatch(score=0.999, text=Bill Gates, offset=0, length=8)], "
-                         "language=English, id=Bill Gates, url=https://en.wikipedia.org/wiki/Bill_Gates, data_source="
-                         "wikipedia)", repr(linked_entity))
+                         "language=English, data_source_entity_id=Bill Gates, "
+                         "url=https://en.wikipedia.org/wiki/Bill_Gates, data_source=wikipedia)", repr(linked_entity))
         self.assertEqual("RecognizeLinkedEntitiesResult(id=1, entities=[LinkedEntity(name=Bill Gates, "
                          "matches=[LinkedEntityMatch(score=0.999, text=Bill Gates, offset=0, length=8), "
                          "LinkedEntityMatch(score=0.999, text=Bill Gates, offset=0, length=8)], language=English, "
-                         "id=Bill Gates, url=https://en.wikipedia.org/wiki/Bill_Gates, data_source=wikipedia)], "
-                         "statistics=TextDocumentStatistics(character_count=14, "
+                         "data_source_entity_id=Bill Gates, url=https://en.wikipedia.org/wiki/Bill_Gates, "
+                         "data_source=wikipedia)], statistics=TextDocumentStatistics(character_count=14, "
                          "transaction_count=18), is_error=False)", repr(recognize_linked_entities_result))
-        self.assertEqual("SentimentScorePerLabel(positive=0.99, neutral=0.05, negative=0.02)",
+        self.assertEqual("SentimentConfidenceScorePerLabel(positive=0.99, neutral=0.05, negative=0.02)",
                          repr(sentiment_confidence_score_per_label))
-        self.assertEqual("SentenceSentiment(sentiment=neutral, sentiment_scores=SentimentScorePerLabel("
+        self.assertEqual("SentenceSentiment(sentiment=neutral, confidence_scores=SentimentConfidenceScorePerLabel("
                          "positive=0.99, neutral=0.05, negative=0.02), offset=0, length=10, warnings="
                          "['sentence was too short to find sentiment'])", repr(sentence_sentiment))
         self.assertEqual("AnalyzeSentimentResult(id=1, sentiment=positive, statistics=TextDocumentStatistics("
-                         "character_count=14, transaction_count=18), sentiment_scores=SentimentScorePerLabel"
+                         "character_count=14, transaction_count=18), confidence_scores=SentimentConfidenceScorePerLabel"
                          "(positive=0.99, neutral=0.05, negative=0.02), sentences=[SentenceSentiment(sentiment=neutral, "
-                         "sentiment_scores=SentimentScorePerLabel(positive=0.99, neutral=0.05, negative=0.02), "
+                         "confidence_scores=SentimentConfidenceScorePerLabel(positive=0.99, neutral=0.05, negative=0.02), "
                          "offset=0, length=10, warnings=['sentence was too short to find sentiment'])], is_error=False)",
                          repr(analyze_sentiment_result))
         self.assertEqual("DocumentError(id=1, error=TextAnalyticsError(code=invalidRequest, "